refactor: move HangfireOptions to Anela.Heblo.Xcc with new perf tunables
